Implemented deleted objects in Hibernate.
We now stored all documents on the same table, even deleted
ones. An extra field was added "deleted", which will
be true in case the document marks a deleted one.

We try NOT to use SQL or depend on the underlying database,
and use the Java based interfaces whenever possible.